requirements.txt
APScheduler==3.7.0
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-apscheduler==0.5.2
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
idna==2.10
PyJWT==2.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.4
